requirements.txt
Flask==1.1.2
scipy==1.4.1
Jinja2==2.11.2
Keras==2.3.1
pandas==1.1.0
numpy>=1.9.2
Pillow==7.2.0
tensorflow==2.1.0